chainspec: test showing parity chainspec invalid (frontier)
This test shows that the Parity chainspec converter function
is limited to (very)late forks only, because it the fields hexutil.Uint64
cannot be nil.
If the fields cannot be nil, then the values cannot be unset, and so the
values MUST be implemented (even when wanted config, eg. Frontier)
does not implement them (eg. EIP150).