Makefile,tests,build,chainspecs: DifficultyTest generator
Implements difficulty test generation using ref:new
pattern, like StateTest generator.
Given a pair of chainconfigurations, the reference (existing)
config is used to provide test parameters for new
config.

In this case, if the new config dne as a JSON chainspec
file, it will be written.

This implementation is kind of hacky, requiring
a uniq-ifiying shell script to dedupe the resulting
difficulty NDJSON test file.