chainspec: fix Istanbul tests/ - set EIP212/3 if not via 1108
Parity chainspec JSON files use a map for builtin pricing.
The keys are activation blocks numbers.
Sometimes, the map has multiple entries, with duplicate values.
This causes Go marshaling to think there is only one entry (last one).
This left Istanbul configurations parsed from this file istanbul_test.go
without configured EIP212 and EIP213 builtins.

This addition adds the missing logic to support the case
that EIP1108 is a _modification_ of 212/3, and so if it's set,
even as the fake-only entry in the map, then 212/3 must be set
also.

This should only happen on test configs, note.