params,consensus/misc: implement canonical-fork hash require interface
EIP150 sees a required fork hash feature.
This is, strictly speaking, unnecessary. But EIP150 was a protocol
fix because ddos and they wanted things to move along quickly. This
is a way to allow clients to immediately decide correctness of any
canonical block.

Multigeth implements the feature with a map, so that this
can be applied for any blocks.