consensus,params/types: fill out ForkCanonHash method interfaces
This idea is derived from EIP150 block/hash implementation.

It is not strictly necessary for these supported clients in this
case; it was a convenience to enable faster fork detection and
participation in light of the DDoS to which this feature was a reaction.

Only multigeth supports a map for this pattern, which may, but hopefully not ;)
be useful in the future.

Signed-off-by: meows <b5c6@protonmail.com>