params/types/ctypes: fix SetValueTotalForHeight logic
This is a simple algorith that I had trouble with.
One of those my brain kind of hurts because I've overthought
this. One of those this should be sooo simple, how
can I not get this straight?

The trick is figuring out exactly what it should do.
This algorithm should set the bomb delay sum for a given
height, translating between differential values and total
values. The first should sum up to and including a block
height to yield a corresponding total value. The algorithm
does this, but in reverse: finding the differential values
for a differential map given a total value at a height.

Signed-off-by: meows <b5c6@protonmail.com>