eth/downloader: refactor common ancestor methods
Now, the spanSearch algorithm is only used if local
head is within MaxHeaderFetch distance of remote.

Span searches with a remote with a distant head don't
make sense.

The only thing this changes for the 'API' is that
the log lines are now Debug instead of Trace
for the different ancestor search algorithms.

Signed-off-by: meows <b5c6@protonmail.com>