params/parity.json.d/classic_phoenix_difficulty_test.json,params/parity.json.d/eip2384_difficulty_test.json,tests,tests/testdata: generate difficulty tests for eip2384 and etc_phoenix
Refactor logic for writing the test config parity-style
files. They weren't getting written.
It is still pretty ugly.

Includes new difficulty test configs for eip2384 and
the associated phoenix configs, though keep in mind
that although the eip2384 is used as the associated seed
tests for phoenix, they are not 1:1 because phoenix
has long since deactivated the difficulty bomb. So the
seeds are just seed numbers and equivalence in difficulty
outcomes should not be expected.

Parity(OpenEthereum) does not have eip2384 test configs
in its upstream remote repo.

Signed-off-by: meows <b5c6@protonmail.com>