vm,vars: install SLOAD constantGas to eip2200
Whether on not EIP2200 specifies SLOAD constant gas is a matter of interpretation and contention.
The document presumes the simultaneity of 1884, but does not do so explicitly.
However, EIP2200's engineering DOES NEED the SLOAD constant gas at 800, whether it specifies it or not.
Please see https://github.com/etclabscore/core-geth/issues/35 as a starting point for reading through some of the history
here.

For all networks supported, and theoretically all networks properly using 2200,
change is a noop; since 2200 MUST be accompanied by 1884.

The reason for this change is to make the code the same as ethereum/go-ethereum,
eliminating recurring merge conflicts that raise eyebrows.

Signed-off-by: meows <b5c6@protonmail.com>