main,core,params,generic,genesisT,multigeth,tests: refactor multigethv0->multigeth
Refactoring done programmatically.

The original intention was to support the many
versions of multi-geth configurations, since they
have changed in big ways in a relatively short time.

However, this is growing increasingly out-of-scope
and we need to limit support to only the latest
version. This removes the versioning tag at the
end of the package name.

Signed-off-by: meows <b5c6@protonmail.com>