rawdb: refactor freezer goroutine to independent function for remote impl
The 'freeze' method was, in the course of this change set,
refactored to permit use by both the remote freezer and the
local freezer implementations.

The risk with this approach is that an unforeseen bug
would cause the apparently-minor necessary modifications
to the logic to negatively impact functioning of the
default builtin FS ancient store implementation.

This refactoring near-duplicates the logic so that the
builtin ancient store freezer.freeze method is now
unmodified and unimpacted by the remote freezer client,
which gets its own dedication goroutine function now.
This reduces the 'risky' surface area of code modification
against the existing code, albeit at the cost of
a function's non-reuse.

Signed-off-by: meows <b5c6@protonmail.com>