core: init ecbp11355 (MESS) implementation
Implements spec documented here:
https://github.com/etclabscore/artificial-finality/blob/master/MESS.md

MESS is an artificial finality mechanism endowing
the blockchain's current segment with so-called 'gravity'
by requiring a competing chain to satisfy an exponentially
increasing relative total difficulty.

Note that is change causes several tests with reorgs
to fail. This will be addressed by implementing
the feature as a chain configuration option.

Signed-off-by: meows <b5c6@protonmail.com>