ethapi: initial typings for block and header responses
The current RPC marshalling responses for
blocks and headers all use map[string]interface{}
which lacks descriptiveness, and when passed thru
the openrpc discover by reflection library, results
in content descriptor types that are not very
descriptive.

Instead of using map[string]interface{}, we want
to try using typed responses that can provide
more useful information to the reflector
and the user.

Date: 2020-11-14 07:44:28-06:00
Signed-off-by: meows <b5c6@protonmail.com>