main: configure datadir with genesis hash too
As noted in the comment on line 310, when
attaching to a remote client it may be that the
API says that chain id is 0 when the client
is not yet synced past the EIP155 activation
block height.

Requesting and switching on the genesis block
hash fixes this issue except for the foundation/classic
corner case (same genesis, different chain ids).

Date: 2020-12-29 05:14:04-06:00
Signed-off-by: meows <b5c6@protonmail.com>