params,params/types/multigeth,tests: fixup tests w/r/t EIP1283/Disable transations
Notably these tests need to work for the matrix
of all configuration schemas (eg. Multigeth) which
is not always super pretty.

The special case for EIP1283 is handled specific to the
troublesome/indicative test for now; it could also
be handled at the Configurator method level
as a general case, but that's going too deep for me right
now.

Date: 2021-04-01 05:50:23-05:00
Signed-off-by: meows <b5c6@protonmail.com>