core: refactor reorg rules w/r/t MESS aka ECBP1100
The replaces a custom reorgData data type
and associated custom (to core-geth) logic
with as much go-ethereum code as possible.

Reorg disallowance via MESS is now
handled in the ForkChoice.ReorgNeeded logic
ONLY.

A new function ForkChoice.CommonAncestor is
introduced as a utility function, and
a simple unit test is provided for it.

Date: 2022-01-24 14:15:29-08:00
Signed-off-by: meows <b5c6@protonmail.com>