go.mod: Update version of go-openrpc-reflect
go-openrpc-reflect imported an earlier version of go-ethereum, which
imported a different version of btcsuite/btcd, which caused tests
to fail in the crypto package.

go-openrpc-reflect already removed their import of go-ethereum in
november 2020, but core geth never updated to the latest version.

This change updates the version of go-openrpc-reflect to resolve
this ambiguous import.