cmd/geth: bump geth console test timeout to 20s
These tests were failing because they were
timing out and getting killed.
I think the --dev mode key creation may account
for the problematic time spend.
This patch just increases the time the test
is willing to wait for its geth instance
to start up and print the logs the test is waiting
for.

Date: 2022-12-06 09:30:16-08:00
Signed-off-by: meows <b5c6@protonmail.com>