params/types/coregeth,tests: fix GrayGlacier difficulty test
Only the testdata-etc suite has this
sub-suite of tests. (Because our tests/testdata
submodule may be outdated?)

Anyways, the chain config for the GrayGlacier fork
was not-existing, and the configurator
method (EIP5133) for core-geth was not configuring
the difficulty bomb delay table right.

ALSO, NOTE the json field name fix
s/eip5133FBloc/eip5133FBlock/

Date: 2022-12-19 09:55:55-08:00
Signed-off-by: meows <b5c6@protonmail.com>