eth/protocols/snap: [PTAL] undefined: fills, undefined: start
the *Syncer.commitHealer method merged weird.

PTAL:
Like the whole chunk was just... old, I guess.
Why? Is this a git thing? Was there something
core-geth modified in there that caused the
ineffectual merge?
Or another chunk in the file that bumped
the git merge algo off by a paragraph?

Date: 2023-02-21 06:05:23-08:00
Signed-off-by: meows <b5c6@protonmail.com>