params/confp,convert_test: +TestCrush_SkipZeroValues; improve Crush logic
The newly revised Crush logic allows you
to use the crusher only on the composeable interface
types that are available for both the target and source
interfaces passed to it.

This, for example, allows to crush a sparse config
into a filled config without erroring because the
sparse (source) config doesnt/wont have a consensus engine
defined in it.

Date: 2023-02-21 10:31:24-08:00
Signed-off-by: meows <b5c6@protonmail.com>